requirements.txt
# ==============================================================================
# 短视频智能插入算力广告 - Python依赖
# ==============================================================================

# ----------------------
# 核心框架
# ----------------------
pydantic>=2.0.0              # 数据模型验证
python-dotenv>=1.0.0         # 环境变量管理
loguru>=0.7.0                # 日志管理
tqdm>=4.66.0                 # 进度条

# ----------------------
# 视频处理
# ----------------------
moviepy>=1.0.3               # 视频编辑和合成
opencv-python>=4.8.0         # 图像处理、关键帧提取
pillow>=10.0.0               # 图像处理辅助

# ----------------------
# 音频处理
# ----------------------
pydub>=0.25.1                # 音频处理
librosa>=0.10.0              # 音频分析

# ----------------------
# AI模型
# ----------------------
openai-whisper>=20231117     # 语音识别（ASR）
openai>=1.0.0                # OpenAI API（LLM）

# ----------------------
# 人声分离
# ----------------------
demucs                       # 人声分离（Meta开源）

# ----------------------
# 人脸检测
# ----------------------
mtcnn>=0.1.1                 # 人脸检测
tensorflow>=2.13.0           # MTCNN依赖

# ----------------------
# HTTP客户端
# ----------------------
requests>=2.31.0             # ComfyUI API调用
httpx>=0.25.0                # 异步HTTP客户端（可选）

# ----------------------
# 工具库
# ----------------------
numpy>=1.24.0                # 数组和矩阵运算
scipy>=1.11.0                # 科学计算

# ----------------------
# 测试
# ----------------------
pytest>=7.4.0                # 单元测试
pytest-cov>=4.1.0            # 测试覆盖率

# ----------------------
# 可选依赖（进阶功能）
# ----------------------
# scenedetect[opencv]        # 场景检测（非文案视频）
# torch>=2.0.0               # PyTorch（如需GPU加速）
# torchvision>=0.15.0
